+++ VueFluxReactive/Internal/Stream.swift
@@ -1,17 +1,17 @@
 import VueFlux
 
 /// A stream that able to both sends and receive values.
-final class Subject<Value>: Subscribable {
+final class Stream<Value> {
     private lazy var observers = ThreadSafe(Storage<(Value) -> Void>())
     
-    /// Subscribe the observer function to be received the values.
+    /// Add the observer function to be received the values.
     ///
     /// - Prameters:
     ///   - observer: A function to be received the values.
     ///
     /// - Returns: A subscription to unsubscribe given observer.
     @discardableResult
-    func subscribe(observer: @escaping (Value) -> Void) -> Subscription {
+    func add(observer: @escaping (Value) -> Void) -> Subscription {
         return observers.modify { observers in
             let key = observers.add(observer)
